myproject/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
gunicorn>=21.2.0
pillow>=10.2.0
